cron/requirements.txt
scrapy==2.4.1
psycopg2-binary
scrapy-selenium
